added crated and edited page

+++ database/migrations/2018_04_13_033438_create_works_table.php
@@ -16,7 +16,7 @@ class CreateWorksTable extends Migration
         Schema::create('works', function (Blueprint $table) {
             $table->increments('id');
             $table->string('topic')->defult('study');
-            $table->integer('hour');
+            $table->double('hour');
             $table->integer('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
